Performance tweaks

+++ v2/simulation/challenge.go
@@ -19,24 +19,24 @@ const (
 func PassedSurvivalCriteria(c *Creature, s *Simulation, challenge ChallengeType) bool {
 	switch challenge {
 	case LeftSurvive:
-		return c.Loc.X < s.Params.WorldWidth/2
+		return float64(c.Loc.X) < s.Params.WorldWidth/2
 
 	case FarLeftSurvive:
-		return c.Loc.X < s.Params.WorldWidth/10
+		return float64(c.Loc.X) < s.Params.WorldWidth/10
 
 	case RightSurvive:
-		return c.Loc.X > s.Params.WorldWidth/2
+		return float64(c.Loc.X) > s.Params.WorldWidth/2
 
 	case Groups:
 		minNeighbours := 4
-		radius := 4.0
-		isBorder := c.Loc.X < 1 || c.Loc.X >= s.Params.WorldWidth-1 ||
-			c.Loc.Y < 1 || c.Loc.Y >= s.Params.WorldHeight-1
+		radius := float32(4.0)
+		isBorder := float64(c.Loc.X) < 1 || float64(c.Loc.X) >= s.Params.WorldWidth-1 ||
+			float64(c.Loc.Y) < 1 || float64(c.Loc.Y) >= s.Params.WorldHeight-1
 		if isBorder {
 			return false
 		}
-		nearEdge := c.Loc.X < 5 || c.Loc.X > s.Params.WorldWidth-6 ||
-			c.Loc.Y < 5 || c.Loc.Y > s.Params.WorldHeight-6
+		nearEdge := float64(c.Loc.X) < 5 || float64(c.Loc.X) > s.Params.WorldWidth-6 ||
+			float64(c.Loc.Y) < 5 || float64(c.Loc.Y) > s.Params.WorldHeight-6
 		if nearEdge {
 			return false
 		}
@@ -44,12 +44,12 @@ func PassedSurvivalCriteria(c *Creature, s *Simulation, challenge ChallengeType)
 		return len(neighbours)-1 >= minNeighbours // -1 to exclude self
 
 	case Center:
-		cx := s.Params.WorldWidth / 2
-		cy := s.Params.WorldHeight / 2
-		radius := 50.0
+		cx := float32(s.Params.WorldWidth / 2)
+		cy := float32(s.Params.WorldHeight / 2)
+		radius := float32(50.0)
 		dx := c.Loc.X - cx
 		dy := c.Loc.Y - cy
-		return math.Sqrt(dx*dx+dy*dy) <= radius
+		return math.Sqrt(float64(dx*dx+dy*dy)) <= float64(radius)
 
 	case AllSurvive:
 		fallthrough
